move enabled_modelmesh_in_dsc do upper conftest
